Fixed KN. I find myself falling for the casted int and double issues in divisions and multiplications...

+++ include/KneserNey.h
@@ -35,7 +35,7 @@ public:
 	void computeFrequencyStats();
 
 private:
-	int n1, n2, n3, n4;
+	double n1, n2, n3, n4;
 
 	/**
 	 * The raw probability is a discounted probability. The n-gram counts are
